# frontend/lib/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DR-OS Control Plane",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project_v1_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Update Project",
        "operationId": "update_project_v1_projects__project_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/projects/{project_id}/datasets/import-public": {
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Import Public Dataset",
        "operationId": "import_public_dataset_v1_projects__project_id__datasets_import_public_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportPublicDatasetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDatasetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/datasets/register-upload": {
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Register Upload Dataset",
        "operationId": "register_upload_dataset_v1_projects__project_id__datasets_register_upload_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUploadDatasetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDatasetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/datasets": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "List Datasets",
        "operationId": "list_datasets_v1_projects__project_id__datasets_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/datasets/{dataset_id}": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Get Dataset",
        "operationId": "get_dataset_v1_projects__project_id__datasets__dataset_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/datasets/{dataset_id}/snapshots": {
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Create Snapshot",
        "operationId": "create_snapshot_v1_projects__project_id__datasets__dataset_id__snapshots_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatasetSnapshotRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDatasetSnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "List Snapshots",
        "operationId": "list_snapshots_v1_projects__project_id__datasets__dataset_id__snapshots_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetSnapshotListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/datasets/{dataset_id}/policy-checks": {
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Run Policy Checks",
        "operationId": "run_policy_checks_v1_projects__project_id__datasets__dataset_id__policy_checks_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetPolicyCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComplianceLevel": {
        "type": "string",
        "enum": [
          "public",
          "internal",
          "clinical"
        ],
        "title": "ComplianceLevel"
      },
...
      "CreateDatasetResponse": {
        "properties": {
          "dataset": {
            "$ref": "#/components/schemas/DatasetRead"
          },
          "snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DatasetSnapshotRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "workflow_instance_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Instance Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dataset"
        ],
        "title": "CreateDatasetResponse"
      },
      "CreateDatasetSnapshotRequest": {
        "properties": {
          "object_uri": {
            "type": "string",
            "minLength": 1,
            "title": "Object Uri"
          },
          "input_hash_sha256": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "title": "Input Hash Sha256"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count"
          },
          "column_schema_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Column Schema Json"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "object_uri",
          "input_hash_sha256"
        ],
        "title": "CreateDatasetSnapshotRequest"
      },
      "CreateDatasetSnapshotResponse": {
        "properties": {
          "snapshot": {
            "$ref": "#/components/schemas/DatasetSnapshotRead"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "snapshot"
        ],
        "title": "CreateDatasetSnapshotResponse"
      },
...
      "DatasetDetailResponse": {
        "properties": {
          "dataset": {
            "$ref": "#/components/schemas/DatasetRead"
          },
          "current_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DatasetSnapshotRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dataset"
        ],
        "title": "DatasetDetailResponse"
      },
      "DatasetListResponse": {
        "properties": {
          "items": {
            "$ref": "#/components/schemas/Page_DatasetRead_"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "DatasetListResponse"
      },
      "DatasetPolicyCheckResponse": {
        "properties": {
          "snapshot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Snapshot Id"
          },
          "phi_scan_status": {
            "type": "string",
            "title": "Phi Scan Status"
          },
          "deid_status": {
            "type": "string",
            "title": "Deid Status"
          },
          "blocking_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blocking Reasons"
          },
          "allowed": {
            "type": "boolean",
            "title": "Allowed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "snapshot_id",
          "phi_scan_status",
          "deid_status",
          "allowed"
        ],
        "title": "DatasetPolicyCheckResponse"
      },
...
      "DatasetSnapshotListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DatasetSnapshotRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "DatasetSnapshotListResponse"
      },
      "DatasetSnapshotRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "snapshot_no": {
            "type": "integer",
            "title": "Snapshot No"
          },
          "object_uri": {
            "type": "string",
            "title": "Object Uri"
          },
          "input_hash_sha256": {
            "type": "string",
            "title": "Input Hash Sha256"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count"
          },
          "column_schema_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Column Schema Json"
          },
          "deid_status": {
            "$ref": "#/components/schemas/DeidStatus"
          },
          "phi_scan_status": {
            "$ref": "#/components/schemas/PhiScanStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "project_id",
          "dataset_id",
          "snapshot_no",
          "object_uri",
          "input_hash_sha256",
          "column_schema_json",
          "deid_status",
          "phi_scan_status",
          "created_at"
        ],
        "title": "DatasetSnapshotRead"
      },
...
      "DeidStatus": {
        "type": "string",
        "enum": [
          "not_required",
          "pending",
          "completed",
          "failed"
        ],
        "title": "DeidStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportPublicDatasetRequest": {
        "properties": {
          "accession": {
            "type": "string",
            "minLength": 1,
            "title": "Accession"
          },
          "source_kind": {
            "type": "string",
            "minLength": 1,
            "title": "Source Kind"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "accession",
          "source_kind"
        ],
        "title": "ImportPublicDatasetRequest"
      },
...
      "ManuscriptRead": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "manuscript_type": {
            "$ref": "#/components/schemas/ManuscriptType"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "state": {
            "$ref": "#/components/schemas/ManuscriptState"
          },
          "current_version_no": {
            "type": "integer",
            "title": "Current Version No"
          },
          "style_profile_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Style Profile Json"
          },
          "target_journal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Journal"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "created_at",
          "id",
          "tenant_id",
          "project_id",
          "manuscript_type",
          "title",
          "state",
          "current_version_no",
          "style_profile_json"
        ],
        "title": "ManuscriptRead"
      },
...
      "Page_DatasetRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DatasetRead"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "page"
        ],
        "title": "Page[DatasetRead]"
      },
...
      "ProjectDetailResponse": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectRead"
          },
          "active_workflows": {
            "items": {
              "$ref": "#/components/schemas/WorkflowInstanceRead"
            },
            "type": "array",
            "title": "Active Workflows"
          },
          "latest_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DatasetSnapshotRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_manuscript": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ManuscriptRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "review_summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Review Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project"
        ],
        "title": "ProjectDetailResponse"
      },
...
      "ProjectRead": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "project_type": {
            "$ref": "#/components/schemas/ProjectType"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectState"
          },
          "compliance_level": {
            "$ref": "#/components/schemas/ComplianceLevel"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "active_manuscript_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Manuscript Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "created_at",
          "id",
          "tenant_id",
          "name",
          "project_type",
          "status",
          "compliance_level",
          "owner_id"
        ],
        "title": "ProjectRead"
      },
...
      "ProjectState": {
        "type": "string",
        "enum": [
          "draft",
          "running",
          "review_required",
          "approved",
          "archived"
        ],
        "title": "ProjectState"
      },
      "ProjectType": {
        "type": "string",
        "enum": [
          "public_omics",
          "clinical_retrospective",
          "case_report",
          "grant"
        ],
        "title": "ProjectType"
      },
      "RegisterUploadDatasetRequest": {
        "properties": {
          "file_ref": {
            "type": "string",
            "minLength": 1,
            "title": "File Ref"
          },
          "display_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Display Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "file_ref",
          "display_name"
        ],
        "title": "RegisterUploadDatasetRequest"
      },
...
      "UpdateProjectRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectState"
              },
              {
                "type": "null"
              }
            ]
          },
          "compliance_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ComplianceLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_manuscript_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Manuscript Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UpdateProjectRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkflowBackend": {
        "type": "string",
        "enum": [
          "queue_workers",
          "temporal"
        ],
        "title": "WorkflowBackend"
      },
...
      "WorkflowInstanceRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "workflow_type": {
            "type": "string",
            "title": "Workflow Type"
          },
          "state": {
            "$ref": "#/components/schemas/WorkflowState"
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "parent_workflow_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Workflow Id"
          },
          "started_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started By"
          },
          "runtime_backend": {
            "$ref": "#/components/schemas/WorkflowBackend"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "project_id",
          "workflow_type",
          "state",
          "runtime_backend",
          "started_at"
        ],
        "title": "WorkflowInstanceRead"
      },
...
      "WorkflowState": {
        "type": "string",
        "enum": [
          "created",
          "retrieving",
          "retrieved",
          "structuring",
          "structured",
          "analyzing",
          "analyzed",
          "asserting",
          "asserted",
          "writing",
          "verifying",
          "approved",
          "blocked",
          "needs_human",
          "exported",
          "failed"
        ],
        "title": "WorkflowState"
      },
...
    }
  }
}